- Fix Inconsistency Model Tone Analyzer Service - Fix Aggregate Exception - Fix Unit Test
